pyproject.toml
[tool.poetry]
name = "bttrs"
version = "0.1.0"
description = ""
authors = ["Ruslan Mamleev <ruslan@mircod.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.2.5"
django-taggit = "^1.5.1"
loguru = "^0.5.3"
django-timezone-field = "^4.2.1"
django-3-jet = "^1.0.8"
django-axes = "^5.20.0"
django-robots = "^4.0"
django-taggit-helpers = "^0.1.4"
django-modeltranslation = "^0.17.3"
django-ckeditor = "^6.1.0"
Pillow = "^8.3.1"
django-stdimage = "^5.3.0"
django-location-field = "^2.1.0"
django-countries = "^7.2.1"
django-colorfield = "^0.4.2"
django-admin-sortable2 = "^1.0"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
flake8 = "^3.9.2"
isort = "^5.9.2"

[tool.black]
line-length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
